RA-08 Внес необходимые изменения
