+++ app/src/main/java/aps/fithom/startandroidapp/ui/category_list/CategoriesListFragment.kt
@@ -5,12 +5,8 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.GridLayoutManager
 import aps.fithom.startandroidapp.data.local.STUB
 import aps.fithom.startandroidapp.databinding.FragmentCategoriesListBinding
-import aps.fithom.startandroidapp.ui.category_list.rv.CategoryListRVAdapter
-
-private const val RV_CATEGORY_LIST_SPAN_COUNT = 2
 
 class CategoriesListFragment : Fragment() {
 
@@ -35,10 +31,7 @@ class CategoriesListFragment : Fragment() {
     private fun initRecycler() {
         val recycler = binding.rvCategory
         val rvAdapter = CategoryListRVAdapter(STUB.getCategories())
-        with(recycler) {
-            layoutManager = GridLayoutManager(requireContext(), RV_CATEGORY_LIST_SPAN_COUNT)
-            adapter = rvAdapter
-        }
+        recycler.adapter = rvAdapter
     }
 
     override fun onDestroyView() {
