+++ app/src/main/res/layout/item_category.xml
@@ -24,7 +24,6 @@
             android:layout_height="130dp"
             android:scaleType="centerCrop"
             tools:src="@drawable/bcg_categories"
-            android:contentDescription="@string/category_logo"
             />
 
         <TextView
