+++ app/src/main/res/values/strings.xml
@@ -4,4 +4,5 @@
     <string name="btn_favorite_text">favorite</string>
     <string name="category_logo">Category logo</string>
     <string name="category_list">Category list</string>
+    <string name="category_image_content_description">%1$s сategory image</string>
 </resources>
